addressed comments from pull request, also added additional outputs to translation to allow additional confidence measures
